move files to new dir
